Added quotes around ${CLICK_COMMAND}
It these quotes are missing, we run into an expansion issue and the
option --interactive is interpreted as an unknown gnome-terminal option.
With the quotes the whole command 'td --interactive' is appended to the
-e gnome-terminal option.

Result of this issue is that mouse clicks on the applet seem to have not
effect.