Replaced option -e with --
This commit replaces the -e option with -- to get rid of the following
warning and to be prepared for future updates of gnome-terminal.

' # Option “-e” is deprecated and might be removed in a later version of gnome-terminal. '
' #Use “-- ” to terminate the options and put the command line to execute after it.      '